requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
feedparser>=6.0.10
pyyaml>=6.0.1
requests>=2.31.0
fastapi>=0.110.0
uvicorn>=0.29.0
